% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{cum.sum.scalar}
\alias{cum.sum.scalar}
\title{cum.sum.scalar}
\usage{
cum.sum.scalar(yc, xc)
}
\arguments{
\item{yc}{y vector}

\item{xc}{x vector}
}
\value{
a list
}
\description{
Compute cumulative sums.
}
