% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{cum.sum.scalar1}
\alias{cum.sum.scalar1}
\title{cum.sum.scalar1}
\usage{
cum.sum.scalar1(xc)
}
\arguments{
\item{xc}{x vector}

\item{yc}{y vector}
}
\value{
a list
}
\description{
TODO
}
