% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{cum.sum.scalar2}
\alias{cum.sum.scalar2}
\title{cum.sum.scalar2}
\usage{
cum.sum.scalar2(yc, xc)
}
\arguments{
\item{yc}{y vector}

\item{xc}{x vector}
}
\value{
a list
}
\description{
TODO
}
