% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{DC.Simul.SL}
\alias{DC.Simul.SL}
\title{DC.Simul.SL}
\usage{
DC.Simul.SL(y, X, cdataL, ID, overlap)
}
\arguments{
\item{y}{The input data(as a vector)}

\item{cdataL}{Pre-defined cdata list which is from List.C.Data(DMatrix,MR)}

\item{ID}{Indices for potential centroids}

\item{overlap}{Boolean which is TRUE for overlapping clusters / FALSE for non-overlapping clusters}

\item{x}{The input data(as a matrix)}
}
\value{
List of most likely clusters
}
\description{
Detect the Cluster in the Simple Linear Regression for given potential centroids.
}
