% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{DC.Simul.SL2}
\alias{DC.Simul.SL2}
\title{DC.Simul.SL2}
\usage{
DC.Simul.SL2(y, X, cdataL, sum_x1, sum_x2, cs.xxL, ID, overlap)
}
\arguments{
\item{y}{The input data(as a vector)}

\item{X}{The input data(as a matrix)}

\item{cdataL}{Pre-defined cdata list which is from List.C.Data(DMatrix,MR)}

\item{sum_x1}{Pre-computed \eqn{\sum(x_i)}}

\item{sum_x2}{Pre-computed \eqn{\sum(x_i^2)}}

\item{cs.xxL}{Pre-computed cs.xx list which is from \code{cum.sum.scalar1(xc)}}

\item{ID}{Indices for potential centroids}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}
}
\value{
List of most likely clusters
}
\description{
Detect the cluster in the simple linear regression for given potential centroids via simultaneous detection
}
