% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{DC.TStg2.SL}
\alias{DC.TStg2.SL}
\title{DC.TStg2.SL}
\usage{
DC.TStg2.SL(y, X, cdataL, ID, overlap)
}
\arguments{
\item{y}{The input data(as a vector)}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}}

\item{ID}{Indices for potential centroids}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}

\item{x}{The input data(as a matrix), N-by-2 matrix}
}
\value{
List of most likely clusters
}
\description{
Detect the cluster in the simple linear regression for given potential centroids via the 2nd stage in Two-stage detection: the same slope but different intercept.
}
