% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage_SpTm.R
\name{Est.Coeff.TS.ST}
\alias{Est.Coeff.TS.ST}
\title{Est.Coeff.SI.ST}
\usage{
Est.Coeff.TS.ST(yList, XList, Cls1stIndicator, Cls2ndIndicator)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{Cls1stIndicator}{Indicator of clusters identified in stage 1.}

\item{Cls2ndIndicator}{Indicator of clusters identified in stage 2.}
}
\value{
List of coefficients
}
\description{
Estimate coefficients via two-stage detection.
}
