% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous_SpTm.R
\name{Find.Clusters.SI.ST}
\alias{Find.Clusters.SI.ST}
\title{Find.Clusters.SI.ST}
\usage{
Find.Clusters.SI.ST(yList, XList, long, lat, MR, M, overlap, alpha)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{long}{longitude.}

\item{lat}{latitude.}

\item{MR}{Maximum radius.}

\item{M}{Number of simulations.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters.}

\item{alpha}{Significance level}
}
\value{
List of clusters, coefficients, and indicator of cluster membership.
}
\description{
Find multiple cylindrical spatio-temporal clusters sequentially via simultaneous detection. Find and test the cluster in the simple linear regression for given potential centroids.
}
\examples{
data("SE_FakeData_SpTm")
long <- SE_FakeData_SpTm$long; lat <- SE_FakeData_SpTm$lat
yList <-list()
yList[[1]] <- SE_FakeData_SpTm$y1
yList[[2]] <- SE_FakeData_SpTm$y2
yList[[3]] <- SE_FakeData_SpTm$y3
XList <-list()
XList[[1]] <- cbind(rep(1,length(long)), SE_FakeData_SpTm$x1)
XList[[2]] <- cbind(rep(1,length(long)), SE_FakeData_SpTm$x2)
XList[[3]] <- cbind(rep(1,length(long)), SE_FakeData_SpTm$x3)
MR <- 300; M <- 2
Clusters_simST <- Find.Clusters.SI.ST(yList, XList, long, lat,
MR, M, overlap=FALSE, alpha=0.05)
}
