% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous_SpTm.R
\name{Fit.Model.Clusters.ST}
\alias{Fit.Model.Clusters.ST}
\title{Fit.Model.Clusters.ST}
\usage{
Fit.Model.Clusters.ST(yList, XList, long, lat, Clusters)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{long}{longitude.}

\item{lat}{latitude.}

\item{Clusters}{Inherited output from \code{Find.Clusters.SI.ST()}.}
}
\description{
Fit a simple linear regression model with detected clusters.
}
