% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage_SpTm.R
\name{Fit.Model.Clusters.TS.ST}
\alias{Fit.Model.Clusters.TS.ST}
\title{Fit.Model.Clusters.ST}
\usage{
Fit.Model.Clusters.TS.ST(yList, XList, Cls1st, Cls2nd)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{Cls1st}{Output from \code{Find.Clusters.TS.ST1()} function in the 1st Stage.}

\item{Cls2nd}{Output from \code{Find.Clusters.TS.ST2()} function in the 2nd Stage.}
}
\description{
Fit a simple linear regression model with detected clusters.
}
