% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{Fstat.CDL.ID.SL2}
\alias{Fstat.CDL.ID.SL2}
\title{Fstat.CDL.ID.SL2}
\usage{
Fstat.CDL.ID.SL2(y, X, cdataL, ID, overlap)
}
\arguments{
\item{y}{The input data (in vector format).}

\item{X}{The input data (in matrix format), N-by-2 matrix.}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters.}
}
\value{
Most likely cluster (defined by center and radius) and \eqn{\hat{\sigma}^2}.
}
\description{
Find the most likely cluster in the simple linear regression for given potential centroids in the intercept only: the same slope but different intercepts.
}
