% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{Fstat.CDL.ID.SL2.ST}
\alias{Fstat.CDL.ID.SL2.ST}
\title{Fstat.CDL.ID.SL2.ST}
\usage{
Fstat.CDL.ID.SL2.ST(yList, XList, cdataL, ID, overlap)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}
}
\value{
Most likely cluster (defined by center and radius), sum of squared errors 1 (from model which has clusters in intercept only).
}
\description{
Find the spatio-temporal cluster estimate in the simple linear regression for given potential centroids in the intercept only: the same slope but different intercept.
}
