% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{List.C.Data}
\alias{List.C.Data}
\title{List.C.Data}
\usage{
List.C.Data(DMatrix, MR)
}
\arguments{
\item{DMatrix}{Distance matrix}

\item{MR}{Maximum radius.}
}
\value{
Matrix
}
\description{
Construct  a list of all possible clusters where r <= MR for each center.
}
