% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{MLC.CDL.ID.1AOVbi}
\alias{MLC.CDL.ID.1AOVbi}
\title{MLC.CDL.ID.1AOVbi}
\usage{
MLC.CDL.ID.1AOVbi(y, X, cdataL, ID, overlap)
}
\arguments{
\item{y}{The input data (in vector format, 2N-by-1).}

\item{X}{The input data (in matrix format, 2N-by-2).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}
}
\value{
Most likely cluster (defined by center and radius) and \eqn{\hat{\sigma}^2}.
}
\description{
Find the most likely cluster in the one-way ANOVA with a binary covariate for given potential centroids.
}
