% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{MLC.CDL.ID.SL}
\alias{MLC.CDL.ID.SL}
\title{MLC.CDL.ID.SL}
\usage{
MLC.CDL.ID.SL(y, X, cdataL, ID, overlap)
}
\arguments{
\item{y}{The input data (in vector format).}

\item{X}{The input data (in matrix format).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters.}
}
\value{
Most likely cluster (defined by center and radius) and \eqn{\hat{\sigma}^2}.
}
\description{
Find the most likely cluster in the simple linear regression for given centroids.
}
