% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{MLC.CDL.ID.SL.ST}
\alias{MLC.CDL.ID.SL.ST}
\title{MLC.CDL.ID.SL.ST}
\usage{
MLC.CDL.ID.SL.ST(yList, XList, cdataL, ID, overlap)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}
}
\value{
Most likely cluster (defined by center and radius) and SSE (error sum of squares).
}
\description{
Find the spatio-temporal cluster estimate in the simple linear regression for given potential centroids.
}
