% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{MLC.CDL2.ID.SL}
\alias{MLC.CDL2.ID.SL}
\title{MLC.CDL2.ID.SL}
\usage{
MLC.CDL2.ID.SL(y, x, cdataL, sum_x1, sum_x2, cs.xxL, ID, overlap)
}
\arguments{
\item{y}{The input data (in vector format).}

\item{x}{The input data (in vector format).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{sum_x1}{Pre-computed \eqn{\sum x_i}.}

\item{sum_x2}{Pre-computed \eqn{\sum x_i^2}.}

\item{cs.xxL}{Pre-computed \code{cs.xx} list which is from \code{cum.sum.scalar1(xc)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters.}
}
\value{
Most likely cluster (defined by center and radius) and \eqn{\hat{\sigma}^2}.
}
\description{
Find the most likely cluster in the simple linear regression for given potential centroids.
}
