% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{MLC.CDL2.ID.SL.ST}
\alias{MLC.CDL2.ID.SL.ST}
\title{MLC.CDL2.ID.SL.ST}
\usage{
MLC.CDL2.ID.SL.ST(yList, xList, cdataL, sum_x1L, sum_x2L, cs.xxL.L, ID,
  overlap)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{xList}{The input data (as a list of vectors).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{sum_x1L}{List of pre-computed \eqn{\sum x_i}.}

\item{sum_x2L}{List of pre-computed \eqn{\sum x_i^2}.}

\item{cs.xxL.L}{List of pre-computed \code{cs.xx} list which is from \code{cum.sum.scalar1(xc)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}
}
\value{
Most likely cluster (defined by center and radius) and SSE (error sum of squares).
}
\description{
Find the spatio-temporal cluster estimate in the simple linear regression for given potential centroids.
}
