% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_FakeData.R
\docType{data}
\name{SE_FakeData}
\alias{SE_FakeData}
\title{South-Eastern United Status - Fake Incidence Counts}
\format{
An object of class "data.frame". Includes the following variables:
\itemize{
   \item 1) State - Seven states of South-Eastern United States (Alabama, Florida, Georgia, Mississippi, North Carolina, South Carolina, and Tennessee); factor
   \item 2) FIPS - County-level FIPS codes in South-Eastern United States
   \item 3) long - longitude
   \item 4) lat - latitude
   \item 5) x - covariate
   \item 6) y - response
}
}
\usage{
data(SE_FakeData)
}
\description{
Fake dataset of disease incidece in South-Eastern United States. This dataset is used for illustration of the package in the vignette.
}
\examples{
data(SE_FakeData)

}
\keyword{datasets}
