% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_FakeData_SpTm.R
\docType{data}
\name{SE_FakeData_SpTm}
\alias{SE_FakeData_SpTm}
\title{South-Eastern United Status - Fake Incidence Counts across Time and Space}
\format{
An object of class "data.frame". Includes the following variables:
\itemize{
   \item 1) State - Seven states of South-Eastern United States (Alabama, Florida, Georgia, Mississippi, North Carolina, South Carolina, and Tennessee); factor
   \item 2) FIPS - County-level FIPS codes in South-Eastern United States
   \item 3) long - longitude
   \item 4) lat - latitude
   \item 5) x1-x3 - covariates
   \item 6) y1-y3 - responses across three time periods
}
}
\usage{
data(SE_FakeData_SpTm)
}
\description{
Fake dataset of disease incidece in South-Eastern United States with time. This dataset is used for illustration of the spatio-temporal framework in the vignette.
}
\examples{
data(SE_FakeData_SpTm)

}
\keyword{datasets}
