% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{Test.Cluster.TStg2.SL}
\alias{Test.Cluster.TStg2.SL}
\title{Test.Cluster.TStg2.SL}
\usage{
Test.Cluster.TStg2.SL(y, X, cdataL, M, ID, overlap)
}
\arguments{
\item{y}{The input data(as a vector)}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}}

\item{M}{number of simulations}

\item{ID}{Indices for potential centroids}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}

\item{x}{The input data(as a matrix)}
}
\value{
Most likely cluster, maximum F-statistic (of all simulations), and associated p-value.
}
\description{
Find and test the cluster in the simple linear regression for given potential centroids via the 2nd stage in Two-stage detection: the same slope but different intercept.
}
