% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage_SpTm.R
\name{Test.MLC.TS.ID.ST1}
\alias{Test.MLC.TS.ID.ST1}
\title{Test.MLC.TS.ID.ST1}
\usage{
Test.MLC.TS.ID.ST1(yList, XList, cdataL, M, ID, overlap)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{M}{Number of simulations}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}
}
\value{
Most likely cluster, maximum F-statistic (of all simulations), and associated p-value.
}
\description{
Find and test the cylindrical spatio-temporal cluster in the slopes for given ID via two-stage detection (in the 1st Stage).
}
