% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefclust.R
\name{TStg.CDL.ID.SL.ST}
\alias{TStg.CDL.ID.SL.ST}
\title{TStg.CDL.ID.SL.ST}
\usage{
TStg.CDL.ID.SL.ST(yList, XList, cdataL, ID, overlap)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{cdataL}{Pre-defined cdata list which is from \code{List.C.Data(DMatrix,MR)}.}

\item{ID}{Indices for potential centroids.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}

\item{xList}{The input data (as a list of matrices).}
}
\value{
List of most likely clusters. First element contains information for the most likely cluster based on slope. The second element contains information for the most likely cluster based on intercept. For each element (slope and intercept), MLC center and radius, SSE (error sum of squares) for the slope/intercept, sum of squared error (from the model which has the cluster in both the intercept and slope), and max F-statistic for the slope/intercept.
}
\description{
Find the spatio-temporal cluster estimate in the simple linear regression for given potential centroids in the slope (1st Stage) and the intercept (2nd stage).
}
