% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage_SpTm.R
\name{Find.Clusters.TS.ST2}
\alias{Find.Clusters.TS.ST2}
\title{Find.Clusters.TS.ST2}
\usage{
Find.Clusters.TS.ST2(yList, XList, long, lat, MR, M, Cls1st, overlap, alpha)
}
\arguments{
\item{yList}{The input data (as a list of vectors).}

\item{XList}{The input data (as a list of matrices).}

\item{long}{longitude.}

\item{lat}{latitude.}

\item{MR}{Maximum radius.}

\item{M}{Number of simulations.}

\item{Cls1st}{Output from \code{Find.Clusters.TS.ST1()} function in the 1st Stage.}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters.}

\item{alpha}{Significance level}
}
\value{
List of clusters, coefficients, and indicator of cluster membership.
}
\description{
Find multiple (overlapping) cylindrical spatio-temporal clusters sequentially in the intercepts via two-stages detection (in the 2nd stage)
}
\examples{
data("SE_FakeData_SpTm")
long <- SE_FakeData_SpTm$long
lat <- SE_FakeData_SpTm$lat
yList <-list()
yList[[1]] <- SE_FakeData_SpTm$y1
yList[[2]] <- SE_FakeData_SpTm$y2
yList[[3]] <- SE_FakeData_SpTm$y3
XList <-list()
XList[[1]] <- cbind(rep(1,length(long)), SE_FakeData_SpTm$x1)
XList[[2]] <- cbind(rep(1,length(long)), SE_FakeData_SpTm$x2)
XList[[3]] <- cbind(rep(1,length(long)), SE_FakeData_SpTm$x3)
MR <- 300
M <- 2
Clusters_ts1N <- Find.Clusters.TS.ST1(yList, XList, long, lat,
MR, M, overlap=FALSE, alpha=0.05)
Clusters_ts2N <- Find.Clusters.TS.ST2(yList, XList, long, lat,
MR, M, Cls1st=Clusters_ts1N, overlap=FALSE, alpha=0.05)

### With the Bonferroni correction
Clusters_ts1B <- Find.Clusters.TS.ST1(yList, XList, long, lat,
MR, M, overlap=FALSE, alpha=(0.05/2))
Clusters_ts2B <- Find.Clusters.TS.ST2(yList, XList, long, lat,
MR, M, Cls1st=Clusters_ts1B, overlap=FALSE, alpha=(0.05/2))
}
