% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{Find.Clusters.TStg1}
\alias{Find.Clusters.TStg1}
\title{Find.Clusters.TStg1}
\usage{
Find.Clusters.TStg1(y, X, long, lat, MR, M, overlap, alpha)
}
\arguments{
\item{y}{The input data(as a vector)}

\item{X}{The input data(as a matrix)}

\item{long}{longitude}

\item{lat}{latitude}

\item{MR}{Maximum radius}

\item{M}{number of simulations}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}

\item{alpha}{significance level}
}
\value{
list of cluster, coefficient
}
\description{
Find multiple clusters sequentially via simulataneous detection.
Find and test the cluster in the simple linear regression for given potential centroids via
the 1st stage in Two-stage detection: different slope and different intercept.
}
