% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{Find.Clusters.TStg2}
\alias{Find.Clusters.TStg2}
\title{Find.Clusters.TStg2}
\usage{
Find.Clusters.TStg2(y, X, long, lat, MR, M, Cls1st, overlap, alpha)
}
\arguments{
\item{y}{The input data(as a vector)}

\item{X}{The input data(as a matrix)}

\item{long}{longitude}

\item{lat}{latitude}

\item{MR}{Maximum radius}

\item{M}{number of simulations}

\item{Cls1st}{the output from \code{Find.Clusters.TStg1} in the 1st-Stage}

\item{overlap}{Boolean which is \code{TRUE} for overlapping clusters / \code{FALSE} for non-overlapping clusters}

\item{alpha}{significance level}
}
\value{
list of cluster, coefficient, and indicator of cluster membership.
}
\description{
Find multiple clusters sequentially via Simulataneous detection.
Find and test the cluster in the simple linear regression for given potential centroids via the 2nd stage in Two-stage detection: the same slope but different intercept.
}
\examples{
data("SE_FakeData")
y <- SE_FakeData$y
X <- cbind(rep(1,length(y)), SE_FakeData$x)
long <- SE_FakeData$long; lat <- SE_FakeData$lat
MR <- 300; M <- 2
Clusters_ts1 <- Find.Clusters.TStg1(y, X, long, lat, MR, M, overlap=TRUE, alpha=0.05)
Clusters_ts2 <- Find.Clusters.TStg2(y, X, long, lat, MR, M, Clusters_ts1, overlap=TRUE, alpha=0.05)
}
